+++ src/main/java/com/atlauncher/dbus/DBusUtils.java
@@ -39,6 +39,10 @@ import org.freedesktop.dbus.types.Variant;
 
 public class DBusUtils {
     public static File[] selectFiles() {
+        return selectFiles(false);
+    }
+
+    public static File[] selectFiles(Boolean directory) {
         try {
             DBusConnection bus = DBusConnection.getConnection(DBusConnection.DBusBusType.SESSION);
 
@@ -82,7 +86,7 @@ public class DBusUtils {
                     "/org/freedesktop/portal/desktop", FileChooserInterface.class);
 
             Map<String, Variant> options = new HashMap<>();
-            options.put("directory", new Variant(Boolean.FALSE));
+            options.put("directory", new Variant(directory));
             options.put("handle_token", new Variant(token));
 
             fileChooserInterface.OpenFile("", "Pick File", options);